+++ Coding Tracker/squre.cs
@@ -9,15 +9,15 @@ namespace Coding_Tracker
     internal class squre : ISqure
     {
         /* write functions to calculate area and perimeter for squre*/
-        public double Calc_Area(float length, float width)
+        public float Calc_Area(float length, float width)
         {
-            double area = length * width;
+            float area = length * width;
             return area;
         }
 
-        public double Calc_Perimeter(float length, float width)
+        public float Calc_Perimeter(float length, float width)
         {
-            double perimeter = length + width;
+            float perimeter = length + width;
             return perimeter;
         }
     }
